Refactor taxonomies list into a component
